Specify the query result output location
Apparently the bucket that I specified when I created the database was
the location for the result of the _query creating the database_, not
the query result location for queries run against that database. I
believe the query result location for queries run against a database
lives on the workgroup.