Give the athena group read access to data buckets
I tried following these instructions:

https://aws.amazon.com/premiumsupport/knowledge-center/athena-analyze-access-logs/

When I got to the step:

> 4.     Under Tables in the left pane, choose Preview table from the
> menu button that's next to the table name. The data from the ALB access
> logs is visible in the Results window.

I got the error:

> Your query has the following error(s):

> com.amazonaws.services.s3.model.AmazonS3Exception: Access Denied
> (Service: Amazon S3; Status Code: 403; Error Code: AccessDenied; Request
> ID...

I believe this is because the user athena-user-1 (who I was logged in as
when I tried this) does not have read access on the buckets where the
actual data is stored.

This view is supported by this:

https://aws.amazon.com/premiumsupport/knowledge-center/access-denied-athena/

This change grants the athena users group read access to the specific
buckets where I have ALB access log data.

My template for read-only access to buckets is the AWS-managed policy
AmazonS3ReadOnlyAccess.